+++ src/main/java/fr/hephaisto/ranking/calculators/Calculator.java
@@ -3,5 +3,5 @@ package fr.hephaisto.ranking.calculators;
 import com.massivecraft.factions.entity.Faction;
 
 public interface Calculator {
-    int compute(Faction faction);
+    double compute(Faction faction);
 }
